[Change] Work on a better (modular) precompiling system
pymdx now has a seperate precompiling method so that one may
use pymdx to either automatically set loop points, etc. or set them
manually.
At this point setting exporting to manual mode is not implemented
yet, but would take a very small effort.

Code still needs to be cleaned up and needs to have better variable
naming. The comments in the code needs to be updated as well
(repeat commands in particular).