[Change] pymdx tone object is not dependent on libpyfmtool anymore
The dependency of libpyfmtool for pymdx's tone object has been removed.
It has been removed as the library is very imcomplete at the moment of writing
and has therefor received its own small definition of an OPM tone object.

This does mean future optimizations to tone objects in compiling have to be done
manually in the font-end of a compiler, atleast until libpyfmtool returns as depedency.